requirements.txt
sanic==0.7.0
requests==2.19.1
python-dotenv==0.9.1
